perf: 实验③④⑤——探测缓存 + uint8 比较 + hybrid 批量交付/多轮校准
- video_utils: nvdec_available/tensorrt_available 加 lru_cache（批量多集
  避免重复打开视频/glob 扫描；实测批量 3 次省 0.05s，零风险保留）
- extractor._segments_similar: float32 双数组改 int16 精确差（省内存、
  与 GPU sim_pair 整数语义一致；实测与 float32 性能持平、段数/文本逐位一致）
- hybrid_decode: get_batch/next_roi 改 _pop_frames 批量交付（一次锁取
  同片连续帧，减少锁次数）；校准支持 HYBRID_CALIB_ROUNDS 多轮取中位数
  （默认 1 轮保持现状——实测 3 轮成本 0.68s > 分界精度收益，保留开关）
- 实测：实验②（OCR 会话提前到校准前）真实收益 ~0.76s（-19%，
  引擎加载 0.4s 从墙钟消失，decode 主导时被完全隐藏）；段数/文本一致

+++ video_ocr_engine/extractor.py
@@ -376,7 +376,10 @@ class FieldExtractor(_GpuPipelineMixin):
             b = _text_sep_gray(b, 'binary', th=self._bin_thresh)
         if a is None or b is None or a.shape != b.shape:
             return False
-        diff = np.abs(a.astype(np.float32) - b.astype(np.float32))
+        # int16 精确差：避免 float32 双全帧临时数组（a/b 为 uint8 灰度或
+        # float32 分离图，255.0/0.0 转 int16 精确）。与 GPU sim_pair 的
+        # 整数精确累加一致（阈值处仅 float32 末位舍入差异，文档已承认）。
+        diff = np.abs(a.astype(np.int16) - b.astype(np.int16))
         if float(diff.mean()) > self._merge_similar_threshold:
             return False
         changed = int(np.sum(diff > 10))
